consumer/requirements.txt
kafka-python==2.1.2
pymongo==4.5.0 
python-dotenv==1.0.1  
